chore: rename utils/process.ts to process-control.ts
Avoids ambiguity with Node's built-in process module.
